perf(frontend): restore code splitting, split create-quest, portal dialog
- Restore React.lazy() route-level code splitting in routes.tsx (#916)
- Add static route→import() prefetch map in PrefetchLink.tsx
- Split create-quest into per-step lazy chunks: step1, step2, step3 (#917)
- Portal-render TransactionConfirmDialog via createPortal (#918)

Fixes #914, #916, #917, #918
