+++ frontend/src/routes.tsx
@@ -1,4 +1,4 @@
-import { Suspense, type ReactNode } from "react"
+import { Suspense, lazy, type ReactNode } from "react"
 import { createBrowserRouter } from "react-router-dom"
 import { AppShell } from "@/components/app-shell"
 import { WalletRequiredRoute } from "@/components/wallet-required-route"
@@ -6,21 +6,15 @@ import { PageSkeleton } from "@/components/page-skeleton"
 import { QuestRedirect } from "@/components/workspace-redirect"
 import { ErrorBoundary } from "@/components/error-boundary"
 
-// Static page imports. We previously used React.lazy() with dynamic imports
-// here so each page would code-split into its own chunk, but Vite 8 / Rolldown
-// emits those dynamic imports as literal runtime fetches against the asset
-// folder without ever emitting a matching chunk file (e.g. /assets/pages/landing
-// 404s in production). Until that toolchain bug is resolved, ship one bundle
-// so navigation actually works. Bundle size is unchanged — Rolldown was
-// inlining all routes into the main chunk anyway.
-import { Landing } from "./pages/landing"
-import { Dashboard } from "./pages/dashboard"
-import { QuestView } from "./pages/quest"
-import { Profile } from "./pages/profile"
-import { NotFound } from "./pages/not-found"
-import { CreateQuest } from "./pages/create-quest"
-import { Leaderboard } from "./pages/leaderboard"
-import { CreatorProfile } from "./pages/creator"
+// Lazy-loaded page chunks — each route downloads only when navigated to.
+const Landing = lazy(() => import("./pages/landing").then(m => ({ default: m.Landing })))
+const Dashboard = lazy(() => import("./pages/dashboard").then(m => ({ default: m.Dashboard })))
+const QuestView = lazy(() => import("./pages/quest").then(m => ({ default: m.QuestView })))
+const Profile = lazy(() => import("./pages/profile").then(m => ({ default: m.Profile })))
+const NotFound = lazy(() => import("./pages/not-found").then(m => ({ default: m.NotFound })))
+const CreateQuest = lazy(() => import("./pages/create-quest").then(m => ({ default: m.CreateQuest })))
+const Leaderboard = lazy(() => import("./pages/leaderboard").then(m => ({ default: m.Leaderboard })))
+const CreatorProfile = lazy(() => import("./pages/creator").then(m => ({ default: m.CreatorProfile })))
 
 function RouteShell({ children, label }: { children: ReactNode; label: string }) {
   return (
